+++ global.R
@@ -115,13 +115,17 @@ CALIPSEO_SHINY_ENV <- new.env()
 #utilities
 #---------------------------------------------------------------------------------------
 #core R script utils
+source("assets/core/utils.R") #to load loggers
 core_assets <- list.files("assets/core", pattern = ".R", full.names = T)
 for(core_asset in core_assets) source(core_asset)
 
 #country R script utils
-country_assets <- list.files(path = file.path("./assets/country", appConfig$country_profile$iso3), 
+country_assets <- list.files(path = file.path("../calipseo-data/country", appConfig$country_profile$iso3), 
                              pattern = ".R", recursive = TRUE, full.names = TRUE)
-for(country_asset in country_assets){ source(country_asset) }
+for(country_asset in country_assets){
+  INFO("Loading country R script '%s'", country_asset)
+  source(country_asset) 
+}
 
 #country profile
 #---------------------------------------------------------------------------------------